refactor: moved event types to event.ts
